move template parsing to request time
This change allows setting the `expand` options in both the
`Excon.new` method, as well as the `Excon::Connection#request`
method.

In practice, this means the following is now possible:

    conn = Excon.new('https://example.com/{uid}')
    conn.get(expand: { uid: 'hello' })

Before, the `expand` option had to be passed into the `new` method
together with the url itself. Now, it can be provided at a later
point in time.

Also, this allows merging of template variables:

    conn = Excon.new('https://example.com/{uid}{?message}', expand: { uid: 'hello' })
    conn.get(expand: { uid: 'world', message: 'universe' })